arsenic_awareness_chatbot/requirements.txt
# Arsenic Awareness Chatbot - Dependencies (Ollama, no API key)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
langdetect>=1.0.9
pydantic>=2.0.0
